Add workspace data and new dependencies for iOS project
This commit includes the creation of a new workspace data for the iOS Todoer project. It also introduces new dependencies specified in the `Gemfile.lock` and `Podfile.lock` ensuring the project has the necessary packages for its functionality and build process. This setup is crucial step for the development and successful build of the iOS application.