Add RefreshControl and refactor layout
Implemented a RefreshControl component that will be used in all parts of the application where refreshing of content is required. This improves the user experience by providing visual feedback when content is being loaded. Additionally, the layout of the todos list and create todo feature were refactored for better component organization and code readability.